@@ -4,35 +4,58 @@
   "description": "",
   "bootstrapVersion": 3,
   "styleLoaders": [],
-  "main": "main.js",
+  "main": "./release/cjs/index.js",
+  "module": "./source/index.ts",
+  "types": "./source/index.ts",
   "scripts": {
+    "webpack": "webpack --mode development",
+	"auto": "tsc -watch - tsconfig.json",
+    "build": "yarn build:esm && yarn build:cjs",
+    "build:esm": "tsc",
+    "build:cjs": "tsc --module commonjs --outDir release/cjs",
+    "tsc2": "tsc && tsc -m es6 --outDir release && webpack --mode development",
+    "tsc": "tsc",
+    "tsc3": "tsc --emitDeclarationOnly",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack-dev-server --mode development",
-    "build": "webpack --mode development",
     "production": "webpack --mode production"
   },
   "repository": {
     "type": "git"
   },
+  "files": [
+    "package.json",
+    "tsconfig.json",
+    "release/*"
+  ],
+  "imports": {
+    "@assets/*": "../assets/*",
+    "@styles/*": "../assets/styles/*",
+    "@components/*": "../source/components/*",
+    "@examples/*": "../source/examples/*",
+    "@utility/*": "../source/utility/*"
+  },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.20.0",
-    "css-loader": "^3.2.0",
-    "html-webpack-plugin": "^3.2.0",
-    "node-sass": "^4.13.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "sass-loader": "^8.0.0",
-    "style-loader": "^1.3.0",
-    "ts-loader": "^8.1.0",
-    "typescript": "^4.2.3",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.10"
+    "axios": "^1.4.0",
+    "css-loader": "^6.8.1",
+    "html-webpack-plugin": "^5.5.3",
+    "module-alias": "^2.2.3",
+    "node-sass": "^9.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "sass-loader": "^13.3.2",
+    "style-loader": "^3.3.3",
+    "ts-loader": "^9.4.4",
+    "typescript": "^5.1.6",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4"
   },
   "devDependencies": {
+    "esbuild-loader": "^3.0.1",
     "react-router-dom": "^6.2.1",
-    "webpack-dev-server": "^3.11.2"
+    "webpack-dev-server": "^4.15.1"
   }
 }
