@@ -1,10 +1,10 @@
 import React, { forwardRef } from "react";
 import ReactDOM from "react-dom";
-import { Props } from "@utility/props";
+import { Props } from "../../utility/props";
 
 import "./styles/index.scss"
-import Common from "@utility/common";
-import { ParseFlexParam } from "@components/Row";
+import Common from "../../utility/common";
+import { ParseFlexParam } from "../../components/Row";
 
 export const Column = ( props ) => {
 	let { className, children, style, flex, ...rest } = props;
