@@ -1,6 +1,6 @@
 import React, { useContext, useEffect, useReducer, useState, useRef , useMemo} from "react";
 import ReactDOM from "react-dom";
-import { Props } from "@utility/props";
+import { Props } from "../../utility/props";
 
 export const Eris = ( props ) => {
 	let { className, children, style, ...rest } = props;
